backend/requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
cloudinary==1.33.0
Django==4.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
PyJWT==2.7.0
python-decouple==3.8
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.16
